# handler.go
package main

import (
	"context"
	"encoding/json"
	"errors"
	"math"
	"net/http"
	"strconv"

	mc "github.com/moge800/gomcprotocol"
)

const (
	maxReadCount   = 1024
	maxWriteValues = 1024
	maxWriteBody   = 1 << 20
)

// writeJSON writes v as JSON with the given status code.
func writeJSON(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(v)
}

func writeErr(w http.ResponseWriter, status int, code, msg string) {
	body := map[string]any{"status": status, "error": msg, "code": code}
	writeJSON(w, status, body)
}

func requireMethod(w http.ResponseWriter, r *http.Request, method string) bool {
	if r.Method == method {
		return true
	}
	w.Header().Set("Allow", method)
	writeErr(w, http.StatusMethodNotAllowed, "bad_request", method+" required")
	return false
}

func requireWritable(w http.ResponseWriter, readonly bool) bool {
	if !readonly {
		return true
	}
	writeErr(w, http.StatusForbidden, "forbidden", "operation not allowed in read-only mode")
	return false
}

func requireRemoteEnabled(w http.ResponseWriter, enableRemote bool) bool {
	if enableRemote {
		return true
	}
	writeErr(w, http.StatusForbidden, "forbidden", "remote-control operations are disabled (use -enable-remote to enable)")
	return false
}

func writePLCErr(w http.ResponseWriter, err error) {
	var busy *busyErr
	var connErr *connErrWrap
	var plcErr *mc.MCProtocolError
	switch {
	case errors.As(err, &busy):
		w.Header().Set("Retry-After", "1")
		writeErr(w, http.StatusServiceUnavailable, "busy", err.Error())
	case errors.Is(err, errQueueClosed):
		writeErr(w, http.StatusServiceUnavailable, "queue_closed", err.Error())
	case errors.Is(err, context.Canceled):
		writeErr(w, 499, "request_canceled", "request canceled")
	case errors.Is(err, context.DeadlineExceeded):
		writeErr(w, http.StatusGatewayTimeout, "request_timeout", "request timed out")
	case errors.As(err, &connErr):
		writeErr(w, http.StatusServiceUnavailable, "connection_error", err.Error())
	case errors.As(err, &plcErr):
		body := map[string]any{
			"status":   http.StatusBadGateway,
			"error":    err.Error(),
			"code":     "plc_error",
			"end_code": "0x" + strconv.FormatUint(uint64(plcErr.EndCode), 16),
		}
		writeJSON(w, http.StatusBadGateway, body)
	default:
		writeErr(w, http.StatusBadGateway, "plc_error", err.Error())
	}
}

// GET /openapi.yaml
func handleOpenAPI(spec []byte) http.HandlerFunc {
	return func(w http.ResponseWriter, _ *http.Request) {
		w.Header().Set("Content-Type", "application/yaml; charset=utf-8")
		w.WriteHeader(http.StatusOK)
		_, _ = w.Write(spec)
	}
}

// GET /metrics
func handleMetrics(plc *PLCQueue) http.HandlerFunc {
	return func(w http.ResponseWriter, _ *http.Request) {
		writeJSON(w, http.StatusOK, plc.Metrics())
	}
}

// GET /version
func handleVersion() http.HandlerFunc {
	return func(w http.ResponseWriter, _ *http.Request) {
		writeJSON(w, http.StatusOK, map[string]any{"version": version})
	}
}

// GET /health
func handleHealth(plc *PLCQueue) http.HandlerFunc {
	return func(w http.ResponseWriter, _ *http.Request) {
		connected := plc.IsConnected()

		plcStatus := "ok"
		if !connected {
			plcStatus = "disconnected"
		}
		writeJSON(w, http.StatusOK, map[string]any{
			"plc_status": plcStatus,
			"connected":  connected,
		})
	}
}

// GET /read?addr=D100&count=5
// GET /read?addr=D100&count=5&dword=true  (reads 32-bit double words; word devices only)
func handleRead(plc *PLCQueue) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if !requireMethod(w, r, http.MethodGet) {
			return
		}

		addrStr := r.URL.Query().Get("addr")
		if addrStr == "" {
			writeErr(w, http.StatusBadRequest, "bad_request", "addr is required")
			return
		}
		countStr := r.URL.Query().Get("count")
		count := 1
		if countStr != "" {
			n, err := strconv.Atoi(countStr)
			if err != nil || n < 1 {
				writeErr(w, http.StatusBadRequest, "bad_request", "invalid count")
				return
			}
			count = n
		}
		if count > maxReadCount {
			writeErr(w, http.StatusBadRequest, "bad_request", "count must be "+strconv.Itoa(maxReadCount)+" or less")
			return
		}

		dword := r.URL.Query().Get("dword") == "true"

		if dword && count > maxReadCount/2 {
			writeErr(w, http.StatusBadRequest, "bad_request", "dword count must be "+strconv.Itoa(maxReadCount/2)+" or less")
			return
		}

		da, err := parseAddr(addrStr)
		if err != nil {
			writeErr(w, http.StatusBadRequest, "bad_request", err.Error())
			return
		}

		sint := r.URL.Query().Get("sint") == "true"

		if (dword || sint) && !isWordDevice(da.Device) {
			if dword {
				writeErr(w, http.StatusBadRequest, "bad_request", "dword is only supported for word devices")
			} else {
				writeErr(w, http.StatusBadRequest, "bad_request", "sint is only supported for word devices")
			}
			return
		}

		if isWordDevice(da.Device) {
			readCount := count
			if dword {
				readCount = count * 2
			}
			words, err := plc.ReadWords(r.Context(), da.Device, da.Addr, readCount)
			if err != nil {
				writePLCErr(w, err)
				return
			}
			if dword {
				vals := make([]int64, count)
				for i := range vals {
					lo := uint32(words[i*2])
					hi := uint32(words[i*2+1])
					combined := lo | hi<<16
					if sint {
						vals[i] = int64(int32(combined))
					} else {
						vals[i] = int64(combined)
					}
				}
				writeJSON(w, http.StatusOK, map[string]any{"values": vals})
			} else {
				vals := make([]int, len(words))
				for i, v := range words {
					if sint {
						vals[i] = int(int16(v))
					} else {
						vals[i] = int(v)
					}
				}
				writeJSON(w, http.StatusOK, map[string]any{"values": vals})
			}
		} else {
			bits, err := plc.ReadBits(r.Context(), da.Device, da.Addr, count)
			if err != nil {
				writePLCErr(w, err)
				return
			}
			writeJSON(w, http.StatusOK, map[string]any{"values": bits})
		}
	}
}

// POST /write?addr=D100   body: {"values":[1,2,3]} or {"values":[true,false]}
// POST /write?addr=D100&dword=true   body: {"values":[100000,200000]} (32-bit double words; word devices only)
func handleWrite(plc *PLCQueue, readonly bool) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if !requireMethod(w, r, http.MethodPost) {
			return
		}
		if !requireWritable(w, readonly) {
			return
		}

		addrStr := r.URL.Query().Get("addr")
		if addrStr == "" {
			writeErr(w, http.StatusBadRequest, "bad_request", "addr is required")
			return
		}

		da, err := parseAddr(addrStr)
		if err != nil {
			writeErr(w, http.StatusBadRequest, "bad_request", err.Error())
			return
		}

		dword := r.URL.Query().Get("dword") == "true"
		sint := r.URL.Query().Get("sint") == "true"

		if (dword || sint) && !isWordDevice(da.Device) {
			if dword {
				writeErr(w, http.StatusBadRequest, "bad_request", "dword is only supported for word devices")
			} else {
				writeErr(w, http.StatusBadRequest, "bad_request", "sint is only supported for word devices")
			}
			return
		}

		var body struct {
			Values json.RawMessage `json:"values"`
		}
		r.Body = http.MaxBytesReader(w, r.Body, maxWriteBody)
		if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
			var maxBytesErr *http.MaxBytesError
			if errors.As(err, &maxBytesErr) {
				writeErr(w, http.StatusRequestEntityTooLarge, "bad_request", "body must not be larger than "+strconv.FormatInt(maxBytesErr.Limit, 10)+" bytes")
				return
			}
			writeErr(w, http.StatusBadRequest, "bad_request", "invalid JSON body")
			return
		}
		if len(body.Values) == 0 {
			writeErr(w, http.StatusBadRequest, "bad_request", "values is required")
			return
		}

		if isWordDevice(da.Device) {
			if dword {
				var dvals []int64
				if err := json.Unmarshal(body.Values, &dvals); err != nil {
					writeErr(w, http.StatusBadRequest, "bad_request", "values must be array of integers")
					return
				}
				if len(dvals) == 0 {
					writeErr(w, http.StatusBadRequest, "bad_request", "values must not be empty")
					return
				}
				if len(dvals) > maxWriteValues/2 {
					writeErr(w, http.StatusBadRequest, "bad_request", "values must contain "+strconv.Itoa(maxWriteValues/2)+" items or less")
					return
				}
				words := make([]uint16, len(dvals)*2)
				for i, v := range dvals {
					var u uint32
					if sint {
						if v < math.MinInt32 || v > math.MaxInt32 {
							writeErr(w, http.StatusBadRequest, "bad_request", "sint dword values must be in range -2147483648..2147483647")
							return
						}
						u = uint32(int32(v))
					} else {
						if v < 0 || v > 1<<32-1 {
							writeErr(w, http.StatusBadRequest, "bad_request", "dword values must be in range 0..4294967295")
							return
						}
						u = uint32(v)
					}
					words[i*2] = uint16(u)
					words[i*2+1] = uint16(u >> 16)
				}
				if err := plc.WriteWords(r.Context(), da.Device, da.Addr, words); err != nil {
					writePLCErr(w, err)
					return
				}
			} else if sint {
				var ivals []int
				if err := json.Unmarshal(body.Values, &ivals); err != nil {
					writeErr(w, http.StatusBadRequest, "bad_request", "values must be array of integers")
					return
				}
				if len(ivals) == 0 {
					writeErr(w, http.StatusBadRequest, "bad_request", "values must not be empty")
					return
				}
				if len(ivals) > maxWriteValues {
					writeErr(w, http.StatusBadRequest, "bad_request", "values must contain "+strconv.Itoa(maxWriteValues)+" items or less")
					return
				}
				words := make([]uint16, len(ivals))
				for i, v := range ivals {
					if v < math.MinInt16 || v > math.MaxInt16 {
						writeErr(w, http.StatusBadRequest, "bad_request", "sint word values must be in range -32768..32767")
						return
					}
					words[i] = uint16(int16(v))
				}
				if err := plc.WriteWords(r.Context(), da.Device, da.Addr, words); err != nil {
					writePLCErr(w, err)
					return
				}
			} else {
				var vals []uint16
				if err := json.Unmarshal(body.Values, &vals); err != nil {
					writeErr(w, http.StatusBadRequest, "bad_request", "values must be array of integers")
					return
				}
				if len(vals) == 0 {
					writeErr(w, http.StatusBadRequest, "bad_request", "values must not be empty")
					return
				}
				if len(vals) > maxWriteValues {
					writeErr(w, http.StatusBadRequest, "bad_request", "values must contain "+strconv.Itoa(maxWriteValues)+" items or less")
					return
				}
				if err := plc.WriteWords(r.Context(), da.Device, da.Addr, vals); err != nil {
					writePLCErr(w, err)
					return
				}
			}
		} else {
			var vals []bool
			if err := json.Unmarshal(body.Values, &vals); err != nil {
				writeErr(w, http.StatusBadRequest, "bad_request", "values must be array of booleans")
				return
			}
			if len(vals) == 0 {
				writeErr(w, http.StatusBadRequest, "bad_request", "values must not be empty")
				return
			}
			if len(vals) > maxWriteValues {
				writeErr(w, http.StatusBadRequest, "bad_request", "values must contain "+strconv.Itoa(maxWriteValues)+" items or less")
				return
			}
			if err := plc.WriteBits(r.Context(), da.Device, da.Addr, vals); err != nil {
				writePLCErr(w, err)
				return
			}
		}

		writeJSON(w, http.StatusOK, map[string]any{"ok": true})
	}
}

// POST /remote/run?clear=0&force=false
func handleRemoteRun(plc *PLCQueue, readonly, enableRemote bool) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if !requireMethod(w, r, http.MethodPost) {
			return
		}
		if !requireWritable(w, readonly) {
			return
		}
		if !requireRemoteEnabled(w, enableRemote) {
			return
		}
		clear := 0
		if s := r.URL.Query().Get("clear"); s != "" {
			n, err := strconv.Atoi(s)
			if err != nil || n < 0 || n > 2 {
				writeErr(w, http.StatusBadRequest, "bad_request", "clear must be 0, 1, or 2")
				return
			}
			clear = n
		}
		force := r.URL.Query().Get("force") == "true"

		if err := plc.RemoteRun(r.Context(), clear, force); err != nil {
			writePLCErr(w, err)
			return
		}
		writeJSON(w, http.StatusOK, map[string]any{"ok": true})
	}
}

// POST /remote/stop
func handleRemoteStop(plc *PLCQueue, readonly, enableRemote bool) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if !requireMethod(w, r, http.MethodPost) {
			return
		}
		if !requireWritable(w, readonly) {
			return
		}
		if !requireRemoteEnabled(w, enableRemote) {
			return
		}
		if err := plc.RemoteStop(r.Context()); err != nil {
			writePLCErr(w, err)
			return
		}
		writeJSON(w, http.StatusOK, map[string]any{"ok": true})
	}
}

// POST /remote/pause?force=false
func handleRemotePause(plc *PLCQueue, readonly, enableRemote bool) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if !requireMethod(w, r, http.MethodPost) {
			return
		}
		if !requireWritable(w, readonly) {
			return
		}
		if !requireRemoteEnabled(w, enableRemote) {
			return
		}
		force := r.URL.Query().Get("force") == "true"
		if err := plc.RemotePause(r.Context(), force); err != nil {
			writePLCErr(w, err)
			return
		}
		writeJSON(w, http.StatusOK, map[string]any{"ok": true})
	}
}

// POST /remote/latch-clear
func handleRemoteLatchClear(plc *PLCQueue, readonly, enableRemote bool) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if !requireMethod(w, r, http.MethodPost) {
			return
		}
		if !requireWritable(w, readonly) {
			return
		}
		if !requireRemoteEnabled(w, enableRemote) {
			return
		}
		if err := plc.RemoteLatchClear(r.Context()); err != nil {
			writePLCErr(w, err)
			return
		}
		writeJSON(w, http.StatusOK, map[string]any{"ok": true})
	}
}

// POST /remote/reset
func handleRemoteReset(plc *PLCQueue, readonly, enableRemote bool) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if !requireMethod(w, r, http.MethodPost) {
			return
		}
		if !requireWritable(w, readonly) {
			return
		}
		if !requireRemoteEnabled(w, enableRemote) {
			return
		}
		if err := plc.RemoteReset(r.Context()); err != nil {
			writePLCErr(w, err)
			return
		}
		writeJSON(w, http.StatusOK, map[string]any{"ok": true})
	}
}